Implement Permissions

+++ src/main/java/panda/rainmaker/command/commands/DeleteRoleListCommand.java
@@ -7,6 +7,7 @@ import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEve
 import panda.rainmaker.command.CommandObject;
 import panda.rainmaker.database.models.GuildSettings;
 import panda.rainmaker.util.OptionDataDefs;
+import panda.rainmaker.util.PermissionMap;
 
 import static panda.rainmaker.util.PandaUtil.*;
 import static panda.rainmaker.util.RoleGiverCache.deleteRoleList;
@@ -14,7 +15,7 @@ import static panda.rainmaker.util.RoleGiverCache.deleteRoleList;
 public class DeleteRoleListCommand extends CommandObject {
 
     public DeleteRoleListCommand() {
-        super("delete-role-list", "Deletes a role list.");
+        super("delete-role-list", "Deletes a role list.", true);
         addOptionData(OptionDataDefs.LIST.asOptionData());
     }
 
@@ -25,7 +26,10 @@ public class DeleteRoleListCommand extends CommandObject {
         try {
             Guild guild = getGuildFromSlashCommandEvent(event);
             Member member = getMemberFromSlashCommandEvent(event);
-            memberHasPermission(member, Permission.MANAGE_CHANNEL);
+            PermissionMap permissionCommandPermissions = guildSettings.getPermissionsForCommand(this.getName());
+            boolean hasPermission = memberHasPermission(member, Permission.MANAGE_ROLES, permissionCommandPermissions);
+
+            if (!hasPermission) throw new Exception("Missing permission(s).");
             String listName = getStringFromOption("List name", event.getOption("list"));
             String deleteResult = deleteRoleList(guildSettings, guild, listName);
             passEvent(event, deleteResult);
